Fixed issue with `mobile_device` method being undefined
When sending emails asynchronously (e.g. from DelayedJob), we don't have access
to `ActionController::MobileFu` which is required by the stylesheet rendering.
Therefore emails can't be sent.

Until now we used `asset_path` directly when loding CSS in email, but
`premailer-rails` has some problem with that (with new `sprockets` version), so
we switched to `stylesheet_link_tag`. This causes issue with `mobile-fu` that
adds its own method to `stylesheet_link_tag` call chain.

The solution is to return faster from `mobile-fu` stylesheet renderer if we
don't have access to `ActionController`.